Remove upstream person_detect_model_data as part of the sync.
Since the example code is going to be maintained in the Arduino
repository, we are also keeping the serialized model in the current
repo. As a result, we do not want to sync the model from tflite-micro.

If there is a new model then both the example and the new model can be
updated manually.