Update sync script for the new location of the models in tflite-micro
With https://github.com/tensorflow/tflite-micro/pull/337, the output
tree from tflite-micro project generation has the models in the
tensorflow source tree.

Since we are forking the examples for the Arduino, we are treating the
corresponding models as being part of the example code (and hence need
to be maintained in the Arduino repo).